Works of android port.

Notices.
1. added more native submodules - libpng, libfreetype2, libfontconfig, libexpat (currently, they are on our repo)
2. build directory change : [build directory]/[XXXX] -> [build dir]/[target triples]/[XXXX]
3. added android main
4. uses freeGLUT instead of GLFW on android
4. fixed misc.

Usage: configure --target-triples=arm-linux-androideabi && make
Result: cross-compiled 32-bit so libraries.

+++ src/components/script/dom/bindings/codegen/CodegenRust.py
@@ -1563,7 +1563,7 @@ for (uint32_t i = 0; i < length; ++i) {
             raise TypeError("We don't support nullable enumerated return types "
                             "yet")
         return ("""assert!((%(result)s as uint) < %(strings)s.len());
-let %(resultStr)s: *JSString = JS_NewStringCopyN(cx, ptr::to_unsafe_ptr(&%(strings)s[%(result)s as u32].value[0]) as *i8, %(strings)s[%(result)s as u32].length as u64);
+let %(resultStr)s: *JSString = JS_NewStringCopyN(cx, ptr::to_unsafe_ptr(&%(strings)s[%(result)s as u32].value[0]) as *i8, %(strings)s[%(result)s as u32].length as libc::size_t);
 if %(resultStr)s.is_null() {
   return 0;
 }
